# backend/app/routers/artifacts.py
from pathlib import Path
from uuid import UUID

from fastapi import APIRouter, Depends, HTTPException
from fastapi.responses import FileResponse
from sqlalchemy.orm import Session

from app import models, schemas
from app.auth import AuthContext, require_auth
from app.config import settings
from app.db import get_db

router = APIRouter(prefix="/projects/{project_id}/artifacts", tags=["Artifacts"])

UNAUTHORIZED = {401: {"description": "Token Bearer ausente ou inválido", "model": schemas.ErrorOut}}
NOT_FOUND = {404: {"description": "Projeto, artefato ou arquivo não encontrado", "model": schemas.ErrorOut}}


def _artifacts_dir(project_id: UUID) -> Path:
    d = Path(settings.storage_dir) / "artifacts" / str(project_id)
    d.mkdir(parents=True, exist_ok=True)
    return d


def _project_belongs_to_user(
    db: Session, project_id: UUID, user_id: UUID | None
) -> bool:
    """Projeto existe e (se há usuário no contexto) pertence a ele."""
    q = db.query(models.Project).filter(models.Project.id == project_id)
    if user_id is not None:
        q = q.filter(models.Project.user_id == user_id)
    return q.first() is not None


@router.get(
    "",
    response_model=list[schemas.ArtifactOut],
    summary="List project artifacts",
    response_description="Artefatos (APKs, relatórios…) mais recentes primeiro",
    responses={**UNAUTHORIZED, **NOT_FOUND},
)
def list_artifacts(
    project_id: UUID,
    db: Session = Depends(get_db),
    auth: AuthContext = Depends(require_auth),
):
    if not _project_belongs_to_user(db, project_id, auth.user_id):
        raise HTTPException(status_code=404, detail="Projeto não encontrado")
    return (
        db.query(models.Artifact)
        .filter_by(project_id=project_id)
        .order_by(models.Artifact.id.desc())
        .all()
    )


@router.get(
    "/{artifact_id}",
    summary="Download artifact",
    response_description=(
        "Stream do arquivo (APK) para instalação no celular. "
        "Header `Content-Disposition: attachment` com o nome do arquivo."
    ),
    responses={**UNAUTHORIZED, **NOT_FOUND},
)
def download_artifact(
    project_id: UUID,
    artifact_id: UUID,
    db: Session = Depends(get_db),
    auth: AuthContext = Depends(require_auth),
):
    if not _project_belongs_to_user(db, project_id, auth.user_id):
        raise HTTPException(status_code=404, detail="Projeto não encontrado")
    artifact = (
        db.query(models.Artifact)
        .filter_by(id=artifact_id, project_id=project_id)
        .first()
    )
    if not artifact:
        raise HTTPException(status_code=404, detail="Artefato não encontrado")
    path = _artifacts_dir(project_id) / artifact.filename
    if not path.is_file():
        raise HTTPException(status_code=404, detail="Arquivo não encontrado no storage")
    return FileResponse(
        path,
        filename=artifact.filename,
        media_type=artifact.content_type or "application/octet-stream",
    )