# backend/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Devian Hub API",
    "description": "\nAPI do hub do **Devian** — projetos, chats (Claude Code) e artefatos de build.\n\n**Autenticação:** os endpoints exigem um **Bearer token** com dois formatos:\n\n1. **Token de máquina** (`DEVIAN_API_TOKEN`) — automação/scripts; atua como o admin.\n2. **JWT de sessão** — retornado por `POST /auth/login` após login com a conta\n   Google (via `google_sign_in` no app; sem Firebase). O backend valida o ID\n   token contra as chaves públicas do Google e emite o token de sessão.\n\nOs dados (projetos, chats, mensagens e artefatos) são escopados pelo usuário\nautenticado: cada usuário vê apenas os próprios dados, resolvidos\nautomaticamente a partir do token.\n\nEndpoints públicos: `GET /health` e `POST /auth/login`.\nClique em **Authorize** e cole o token, sem o prefixo `Bearer `.\n",
    "version": "0.6.0"
  },
  "servers": [
    {
      "url": "https://api.agapech.com.br/devian",
      "description": "Produção — via túnel Cloudflare (oracle-hermi)"
    },
    {
      "url": "/devian",
      "description": "Local — backend direto na porta 8088"
    }
  ],
  "paths": {
...
    "/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Login with Google",
        "operationId": "login_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Valida o ID token do Google, cria/atualiza o usuário e devolve o token de sessão + perfil. Público (não exige Bearer).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "401": {
            "description": "Token do Google inválido ou expirado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "403": {
            "description": "Conta não autorizada, inativa ou deletada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "503": {
            "description": "Autenticação não configurada (GOOGLE_CLIENT_ID / SESSION_JWT_SECRET)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "422": {
            "description": "Corpo da requisição inválido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          }
        }
      }
    },
...
    "/projects/{project_id}/artifacts": {
      "get": {
        "tags": [
          "Artifacts"
        ],
        "summary": "List project artifacts",
        "operationId": "list_artifacts_projects__project_id__artifacts_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Artefatos (APKs, relatórios…) mais recentes primeiro",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ArtifactOut"
                  },
                  "title": "Response List Artifacts Projects  Project Id  Artifacts Get"
                }
              }
            }
          },
          "401": {
            "description": "Token Bearer ausente ou inválido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "404": {
            "description": "Projeto, artefato ou arquivo não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/projects/{project_id}/artifacts/{artifact_id}": {
      "get": {
        "tags": [
          "Artifacts"
        ],
        "summary": "Download artifact",
        "operationId": "download_artifact_projects__project_id__artifacts__artifact_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          },
          {
            "name": "artifact_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Artifact Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Stream do arquivo (APK) para instalação no celular. Header `Content-Disposition: attachment` com o nome do arquivo.",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "Token Bearer ausente ou inválido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "404": {
            "description": "Projeto, artefato ou arquivo não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ArtifactOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "size_bytes": {
            "type": "integer",
            "title": "Size Bytes",
            "description": "Size in bytes"
          },
          "content_type": {
            "title": "Content Type",
            "type": "string",
            "nullable": true
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "filename",
          "size_bytes",
          "content_type",
          "created_at"
        ],
        "title": "ArtifactOut",
        "example": {
          "content_type": "application/vnd.android.package-archive",
          "created_at": "2026-08-05T16:10:00Z",
          "filename": "serr-homolog-0.1.0(1).apk",
          "id": "0191f3b2-4c3a-7b00-8000-00000000001e",
          "size_bytes": 29413656
        }
      },
...
      "ErrorOut": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Mensagem de erro (em português)"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "ErrorOut",
        "description": "Resposta de erro padrão da API.",
        "example": {
          "message": "Projeto não encontrado"
        }
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "LoginRequest": {
        "properties": {
          "id_token": {
            "type": "string",
            "title": "Id Token",
            "description": "Google ID token (JWT) returned by google_sign_in (with serverClientId set). The backend validates its signature against Google's public keys — no Firebase involved.",
            "example": "eyJhbGciOiJSUzI1NiIsImtpZCI6IjE..."
          }
        },
        "type": "object",
        "required": [
          "id_token"
        ],
        "title": "LoginRequest",
        "description": "Exchanges a Google ID token for a Devian session token.",
        "example": {
          "id_token": "eyJhbGciOiJSUzI1NiIsImtpZCI6IjE..."
        }
      },
      "LoginResponse": {
        "properties": {
          "token": {
            "type": "string",
            "title": "Token",
            "description": "Devian session JWT — send as `Authorization: Bearer <token>`."
          },
          "user": {
            "$ref": "#/components/schemas/UserOut"
          }
        },
        "type": "object",
        "required": [
          "token",
          "user"
        ],
        "title": "LoginResponse",
        "description": "Session token + user profile.",
        "example": {
          "token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...",
          "user": {
            "created_at": "2026-08-05T16:00:00Z",
            "email": "samuel@agapech.com.br",
            "id": "0191f3b2-4c3a-7b00-8000-0000000000ff",
            "name": "Samuel",
            "picture_url": "https://lh3.googleusercontent.com/a/...",
            "role": "admin",
            "status": "active",
            "updated_at": "2026-08-05T16:00:00Z"
          }
        }
      },
...
      "UserOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "picture_url": {
            "title": "Picture Url",
            "type": "string",
            "nullable": true
          },
          "role": {
            "type": "string",
            "title": "Role",
            "description": "'admin' or 'member'"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "description": "'active', 'blocked' or 'deleted'"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "email",
          "name",
          "picture_url",
          "role",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "UserOut",
        "example": {
          "created_at": "2026-08-05T16:00:00Z",
          "email": "samuel@agapech.com.br",
          "id": "0191f3b2-4c3a-7b00-8000-0000000000ff",
          "name": "Samuel",
          "picture_url": "https://lh3.googleusercontent.com/a/...",
          "role": "admin",
          "status": "active",
          "updated_at": "2026-08-05T16:00:00Z"
        }
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "description": "Token Bearer do Devian — aceita dois formatos:\n1. Token de máquina (DEVIAN_API_TOKEN): atua como o admin (dono).\n2. JWT de sessão: retornado por POST /auth/login após login com Google.\nCole apenas o token, sem o prefixo 'Bearer '.",
        "scheme": "bearer"
      }
    }
  },
...
}